Fix builder helper import cycle for validated config tests
