+++ packages/builder/src/stores/builder/navigation.ts
@@ -1,6 +1,7 @@
 import { derived, get, Writable } from "svelte/store"
 import { API } from "@/api"
-import { appStore, workspaceAppStore } from "@/stores/builder"
+import { appStore } from "./app"
+import { workspaceAppStore } from "./workspaceApps"
 import { DerivedBudiStore } from "../BudiStore"
 import { AppNavigation, AppNavigationLink, UIObject } from "@budibase/types"
 import { notifications } from "@budibase/bbui"
